read hex line by line and decode it
Initial check in for hex file reader module. The module can currently read hex files line by line and decode it. Display the read data if the debugging is on. Creates dummy memory to fill in all the data bytes. There is a segmentation fault that occurs which needs to be fixed.